+++ Packages/com.texelsaur.video/Editor/Inspectors/SyncPlayerInspector.cs
@@ -5,6 +5,7 @@ using UnityEditor.SceneManagement;
 using UdonSharpEditor;
 using System.Collections.Generic;
 using System;
+using VRC.SDKBase;
 
 #if UNITY_2019
 using UnityEditor.Experimental.SceneManagement;
@@ -31,6 +32,7 @@ namespace Texel
         SerializedProperty runBuildHooksProperty;
 
         SerializedProperty defaultUrlProperty;
+        SerializedProperty defaultUrlInterruptibleProperty;
         SerializedProperty defaultLockedProperty;
         SerializedProperty debugLoggingProperty;
         SerializedProperty loopProperty;
@@ -70,6 +72,7 @@ namespace Texel
             runBuildHooksProperty = serializedObject.FindProperty(nameof(SyncPlayer.runBuildHooks));
 
             defaultUrlProperty = serializedObject.FindProperty(nameof(SyncPlayer.defaultUrl));
+            defaultUrlInterruptibleProperty = serializedObject.FindProperty(nameof(SyncPlayer.defaultUrlInterruptible));
             defaultLockedProperty = serializedObject.FindProperty(nameof(SyncPlayer.defaultLocked));
             debugLoggingProperty = serializedObject.FindProperty(nameof(SyncPlayer.debugLogging));
             loopProperty = serializedObject.FindProperty(nameof(SyncPlayer.loop));
@@ -106,7 +109,7 @@ namespace Texel
             GUIStyle boldFoldoutStyle = new GUIStyle(EditorStyles.foldout);
             boldFoldoutStyle.fontStyle = FontStyle.Bold;
 
-            TXLVideoPlayer videoPlayer = (TXLVideoPlayer)serializedObject.targetObject;
+            SyncPlayer videoPlayer = (SyncPlayer)serializedObject.targetObject;
 
             TimeSpan time = DateTime.Now.Subtract(lastValidate);
             if (time.TotalMilliseconds > 1000)
@@ -146,6 +149,13 @@ namespace Texel
             EditorGUILayout.LabelField("URLs & URL Sources", EditorStyles.boldLabel);
 
             EditorGUILayout.PropertyField(defaultUrlProperty, new GUIContent("Default URL", "Optional default URL to play on world load.  If a separate URL Source is also provided, the default URL will play first."));
+            if (videoPlayer.defaultUrl != null && videoPlayer.defaultUrl.Get() != "")
+            {
+                EditorGUI.indentLevel++;
+                EditorGUILayout.PropertyField(defaultUrlInterruptibleProperty, new GUIContent("Interruptible", "Whether the default URL playback can be interrupted by other interrupting sources, like queues."));
+                EditorGUI.indentLevel--;
+            }
+
             if (TXLGUI.DrawObjectFieldWithAdd(sourceManagerProperty, new GUIContent("URL Source Manager", "Manager for queues, playlists, and other URL sources."), new GUIContent("+", "Create new Source Manager")))
                 VideoTxlManager.AddSourceManagerToScene(true);